sm8250-common: Vibrator: Match Effects with kernel environment

+++ vibrator/effect/effect.h
@@ -35,7 +35,7 @@ struct effect_stream {
     uint32_t   	    effect_id;
     uint32_t        length;
     uint32_t        play_rate_hz;
-    const int8_t    *data;
+    const int16_t   *data;
 };
 
 const struct effect_stream *get_effect_stream(uint32_t effect_id);
